Refactored apply functions.
