+++ gettsim/tax_transfer.py
@@ -1,3 +1,4 @@
+from gettsim.apply_tax_funcs import apply_tax_transfer_func
 from gettsim.benefits.alg2 import alg2
 from gettsim.benefits.arbeitslosengeld import ui
 from gettsim.benefits.benefit_checks import benefit_priority
@@ -81,7 +82,7 @@ def tax_transfer(
         "pkv",
     ]
     out_cols = ["svbeit", "rvbeit", "avbeit", "gkvbeit", "pvbeit"]
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df,
         tax_func=soc_ins_contrib,
         level=person,
@@ -110,7 +111,7 @@ def tax_transfer(
     ]
     out_cols = ["elterngeld", "geschw_bonus", "num_mehrlinge", "elternzeit_anspruch"]
 
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df,
         tax_func=elterngeld,
         level=["hid"],
@@ -138,7 +139,7 @@ def tax_transfer(
         "age",
     ]
     out_col = "m_alg1"
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df,
         tax_func=ui,
         level=person,
@@ -154,7 +155,7 @@ def tax_transfer(
     )
     in_cols = ["m_wage", "east", "age", "year", "byear", "exper", "EP"]
     out_col = "pensions_sim"
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df,
         tax_func=pensions,
         level=person,
@@ -206,7 +207,7 @@ def tax_transfer(
         "altfreib",
         "vorsorge",
     ]
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df,
         tax_func=zve,
         level=tax_unit,
@@ -229,7 +230,7 @@ def tax_transfer(
         + [f"tax_{inc}_tu" for inc in e_st_abzuege_params["zve_list"]]
         + ["abgst_tu", "abgst", "soli", "soli_tu"]
     )
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df,
         tax_func=tax_sched,
         level=tax_unit,
@@ -244,7 +245,7 @@ def tax_transfer(
     )
     in_cols = ["age", "w_hours", "ineducation", "m_wage"]
     out_cols = ["kindergeld_basis", "kindergeld_tu_basis"]
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df,
         tax_func=kindergeld,
         level=tax_unit,
@@ -266,7 +267,7 @@ def tax_transfer(
         "kindergeld_hh",
         "kindergeld_tu",
     ]
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df,
         tax_func=favorability_check,
         level=tax_unit,
@@ -287,7 +288,7 @@ def tax_transfer(
         "zveranl",
     ]
     out_col = "uhv"
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df,
         tax_func=uhv,
         level=tax_unit,
@@ -324,7 +325,7 @@ def tax_transfer(
         "handcap_degree",
     ]
     out_cols = ["wohngeld_basis", "wohngeld_basis_hh"]
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df,
         tax_func=wg,
         level=household,
@@ -368,7 +369,7 @@ def tax_transfer(
         "ekanrefrei",
         "ar_alg2_ek",
     ]
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df,
         tax_func=alg2,
         level=household,
@@ -399,7 +400,7 @@ def tax_transfer(
         "uhv",
     ]
     out_cols = ["kiz_temp", "kiz_incrange"]
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df,
         tax_func=kiz,
         level=household,
@@ -425,7 +426,7 @@ def tax_transfer(
         "byear",
     ]
     out_cols = ["kiz", "wohngeld", "m_alg2"]
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df,
         tax_func=benefit_priority,
         level=household,
@@ -455,7 +456,7 @@ def tax_transfer(
         "m_alg2",
     ]
     out_cols = ["dpi_ind", "dpi"]
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df,
         tax_func=disposable_income,
         level=household,
@@ -473,7 +474,7 @@ def tax_transfer(
         "kindergeld",
     ]
     out_col = "gross"
-    df = _apply_tax_transfer_func(
+    df = apply_tax_transfer_func(
         df, tax_func=gross_income, level=household, in_cols=in_cols, out_cols=[out_col]
     )
     required_inputs = [
@@ -558,32 +559,6 @@ def tax_transfer(
         "gross",
     ]
     return df[required_inputs + desired_outputs]
-
-
-def _apply_tax_transfer_func(
-    df, tax_func, level, in_cols, out_cols, func_args=None, func_kwargs=None
-):
-    func_args = [] if func_args is None else func_args
-    func_kwargs = {} if func_kwargs is None else func_kwargs
-
-    df = df.reindex(columns=df.columns.tolist() + out_cols)
-    if len(df.index) == 0:
-        return df
-    else:
-        df.loc[:, in_cols + out_cols] = df.groupby(level)[in_cols + out_cols].apply(
-            _apply_squeeze_function, tax_func, level, func_args, func_kwargs
-        )
-        return df
-
-
-def _apply_squeeze_function(group, tax_func, level, func_args, func_kwargs):
-    if level == ["hid", "tu_id", "pid"]:
-        person = tax_func(group.squeeze(), *func_args, **func_kwargs)
-        for var in person.index:
-            group.loc[:, var] = person[var]
-        return group
-    else:
-        return tax_func(group, *func_args, **func_kwargs)
 
 
 def calculate_tax_and_transfers(
